# openapi.yaml
openapi: 3.0.3
info:
  title: API Local de Venta de Artículos
  description: |
    APIs REST para la gestión de **Productos** y **Órdenes de compra** de un local
    que vende de forma presencial y también recibe pedidos por la Web.

    La solución sigue un estilo arquitectónico de microservicios: los recursos
    `/api/productos` y `/api/ordenes` son atendidos por servicios independientes,
    expuestos aquí bajo una única especificación.

    **Parte 3 - procesamiento por mensajería.** El servicio Publicador (`/api/publicaciones`)
    toma de Órdenes (`GET /api/ordenes?estado=Created`) cada orden lista para procesar y la
    publica en un broker MQTT (Eclipse Mosquitto, tópico `ordenes/procesar`) como JSON
    `{"id", "estado", "fechaCreacion"}`. El servicio
    de Procesamiento (`/api/facturas`, `/api/procesamientos`) recibe el mensaje, verifica que
    la orden no haya sido procesada, valida el stock de cada producto y:
    con stock descuenta el stock, genera la factura y deja la orden en `Ready to Delivery`;
    sin stock la deja en `No Stock`. El cambio de estado se informa con
    `PATCH /api/ordenes/{id}/estado`.
  version: 1.1.0
  contact:
    name: Equipo de desarrollo - Taller IISS (UTEC - ITR Este)
servers:
  - url: http://localhost:8080
    description: Entorno local (Docker Compose - API Gateway / reverse proxy)
  - url: http://localhost:5001
    description: Microservicio de Productos
  - url: http://localhost:5002
    description: Microservicio de Órdenes
...
paths:
...
  /api/ordenes:
    get:
      tags: [Ordenes]
      summary: Lista todas las órdenes de compra
      description: |
        Devuelve los datos de cada orden junto con la lista de productos
        solicitados (identificador y cantidad). Con el parámetro opcional `estado`
        devuelve solo las órdenes en ese estado (el servicio Publicador pide las que
        están en `Created`).
      operationId: listarOrdenes
      parameters:
        - name: estado
          in: query
          required: false
          description: Filtra por estado de la orden
          schema:
            $ref: '#/components/schemas/EstadoOrden'
          example: Created
      responses:
        '200':
          description: Lista de órdenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Orden'
              example:
                - id: 1001
                  email: cliente@email.com
                  direccionEnvio: Av. Italia 3333, Maldonado
                  telefono: '+59899111222'
                  estado: Created
                  fechaCreacion: '2026-06-21T14:30:00-03:00'
                  productos:
                    - productoId: 1
                      cantidad: 2
                    - productoId: 2
                      cantidad: 5
        '400':
          $ref: '#/components/responses/SolicitudInvalida'

    post:
      tags: [Ordenes]
      summary: Registra una nueva orden de compra
      description: |
        Crea la orden en estado `Created`. Si algún producto no existe o el stock
        actual es insuficiente, la orden no se crea y se responde con `409 Conflict`.

        El alta **no descuenta stock**: el servicio Publicador la publica en el broker de
        mensajería y el servicio de Procesamiento descuenta el stock y la deja en
        `Ready to Delivery` (o en `No Stock` si el stock ya no alcanza, p. ej. porque
        otra orden lo consumió antes). El resultado se consulta con `GET /api/ordenes/{id}`.
      operationId: crearOrden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdenCreate'
            example:
              email: cliente@email.com
              direccionEnvio: Av. Italia 3333, Maldonado
              telefono: '+59899111222'
              productos:
                - productoId: 1
                  cantidad: 2
                - productoId: 2
                  cantidad: 5
      responses:
        '201':
          description: Orden creada exitosamente
          headers:
            Location:
              description: URI de la orden creada
              schema:
                type: string
                format: uri
              example: /api/ordenes/1001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenCreada'
              example:
                id: 1001
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
        '409':
          description: |
            Conflicto: alguno de los productos solicitados no existe o no tiene
            stock suficiente para la cantidad pedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                stockInsuficiente:
                  summary: Stock insuficiente
                  value:
                    codigo: 409
                    mensaje: Stock insuficiente para uno o más productos solicitados
                    detalles:
                      - 'Producto 1: stock disponible 3, cantidad solicitada 5'
                productoInexistente:
                  summary: Producto inexistente
                  value:
                    codigo: 409
                    mensaje: Uno o más productos solicitados no existen
                    detalles:
                      - 'No existe el producto con id 99'
  /api/ordenes/{id}:
    parameters:
      - $ref: '#/components/parameters/OrdenId'
    get:
      tags: [Ordenes]
      summary: Obtiene una orden por su identificador
      operationId: obtenerOrden
      responses:
        '200':
          description: Datos de la orden solicitada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orden'
              example:
                id: 1001
                email: cliente@email.com
                direccionEnvio: Av. Italia 3333, Maldonado
                telefono: '+59899111222'
                estado: Created
                fechaCreacion: '2026-06-21T14:30:00-03:00'
                productos:
                  - productoId: 1
                    cantidad: 2
                  - productoId: 2
                    cantidad: 5
        '404':
          $ref: '#/components/responses/OrdenNoEncontrada'
  /api/ordenes/{id}/detalle:
    parameters:
      - $ref: '#/components/parameters/OrdenId'
    get:
      tags: [Ordenes]
      summary: Obtiene una orden con el detalle completo de sus productos
      description: |
        Devuelve los datos de la orden y, por cada producto solicitado, todos los
        datos del producto además de la cantidad pedida.
      operationId: obtenerDetalleOrden
      responses:
        '200':
          description: Orden con el detalle completo de cada producto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenDetalle'
              example:
                id: 1001
                email: cliente@email.com
                direccionEnvio: Av. Italia 3333, Maldonado
                telefono: '+59899111222'
                estado: Created
                fechaCreacion: '2026-06-21T14:30:00-03:00'
                total: 3000.50
                productos:
                  - cantidad: 2
                    subtotal: 2501.00
                    producto:
                      id: 1
                      nombre: Notebook Lenovo ThinkPad
                      descripcion: Notebook Intel i7 16GB RAM
                      precioUnitario: 1250.50
                      stock: 13
                      imagenes:
                        - https://cdn.local/img/thinkpad-1.jpg
                  - cantidad: 5
                    subtotal: 499.50
                    producto:
                      id: 2
                      nombre: Mouse Logitech MX Master 3
                      descripcion: Mouse inalámbrico ergonómico
                      precioUnitario: 99.90
                      stock: 35
                      imagenes:
                        - https://cdn.local/img/mx-master-1.jpg
        '404':
          $ref: '#/components/responses/OrdenNoEncontrada'
  /api/ordenes/{id}/estado:
    parameters:
      - $ref: '#/components/parameters/OrdenId'
    patch:
      tags: [Ordenes]
      summary: Informa el resultado del procesamiento de una orden
      description: |
        Lo invoca el servicio de Procesamiento. Solo una orden en `Created` puede pasar a
        `Ready to Delivery` o `No Stock`. Repetir el estado ya aplicado responde `200`
        (idempotente); pedir otro estado distinto sobre una orden ya resuelta responde `409`.
      operationId: actualizarEstadoOrden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdenEstadoUpdate'
            example:
              estado: Ready to Delivery
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensaje'
              example:
                mensaje: Estado de la orden actualizado a Ready to Delivery
        '400':
          $ref: '#/components/responses/SolicitudInvalida'
        '404':
          $ref: '#/components/responses/OrdenNoEncontrada'
        '409':
          description: La orden ya no está en `Created` y no admite ese cambio de estado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                codigo: 409
                mensaje: "La orden 1 no puede pasar de 'No Stock' a 'Ready to Delivery'"
                detalles:
                  - "Solo una orden en estado 'Created' puede pasar a 'Ready to Delivery' o 'No Stock'"
...
  /api/facturas:
    get:
      tags: [Procesamiento]
      summary: Lista las facturas generadas
      description: |
        Una factura se genera por cada orden que quedó en `Ready to Delivery`. Con el
        parámetro `ordenId` devuelve solo las facturas de esa orden (lista vacía si no tiene).
      operationId: listarFacturas
      parameters:
        - name: ordenId
          in: query
          required: false
          description: Filtra por identificador de orden
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista de facturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Factura'
              example:
                - id: 1
                  ordenId: 1
                  fecha: '2026-06-21T14:30:05-03:00'
                  email: cliente@email.com
                  direccionEnvio: Av. Italia 3333, Maldonado
                  telefono: '+59899111222'
                  items:
                    - productoId: 1
                      descripcion: Notebook Lenovo ThinkPad
                      cantidad: 2
                      precioUnitario: 1250.50
                      subtotal: 2501.00
                    - productoId: 2
                      descripcion: Mouse Logitech MX Master 3
                      cantidad: 5
                      precioUnitario: 99.90
                      subtotal: 499.50
                  total: 3000.50
  /api/facturas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Identificador único de la factura
        schema:
          type: integer
          format: int64
          minimum: 1
        example: 1
    get:
      tags: [Procesamiento]
      summary: Obtiene una factura por su identificador
      operationId: obtenerFactura
      responses:
        '200':
          description: Factura solicitada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Factura'
        '404':
          description: No existe una factura con el identificador indicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                codigo: 404
                mensaje: No existe la factura con id 999
...
components:
  parameters:
...
    OrdenId:
      name: id
      in: path
      required: true
      description: Identificador único de la orden
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1001
  responses:
...
    SolicitudInvalida:
      description: Los datos enviados no son válidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            codigo: 400
            mensaje: Datos de entrada inválidos
            detalles:
              - 'El campo nombre es obligatorio'
              - 'El campo precioUnitario debe ser mayor que 0'
...
    OrdenNoEncontrada:
      description: No existe una orden con el identificador indicado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            codigo: 404
            mensaje: No existe la orden con id 9999
  schemas:
    Producto:
      type: object
      description: Producto del catálogo
      required: [id, nombre, precioUnitario, stock]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Identificador único generado por el sistema
          example: 1
        nombre:
          type: string
          minLength: 1
          maxLength: 100
          example: Notebook Lenovo ThinkPad
        descripcion:
          type: string
          maxLength: 500
          example: Notebook Intel i7 16GB RAM
        precioUnitario:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          example: 1250.50
        stock:
          type: integer
          format: int32
          minimum: 0
          description: Unidades disponibles
          example: 15
        imagenes:
          type: array
          description: Lista de links a imágenes del producto
          items:
            type: string
            format: uri
            maxLength: 500
          example:
            - https://cdn.local/img/thinkpad-1.jpg
            - https://cdn.local/img/thinkpad-2.jpg
...
    LineaOrden:
      type: object
      description: Producto solicitado en una orden y su cantidad
      required: [productoId, cantidad]
      properties:
        productoId:
          type: integer
          format: int64
          example: 1
        cantidad:
          type: integer
          format: int32
          minimum: 1
          example: 2
    EstadoOrden:
      type: string
      description: |
        Estado del pedido. Las órdenes se crean en `Created`; el servicio de
        Procesamiento las pasa a `Ready to Delivery` o `No Stock`.
      enum:
        - Created
        - Ready to Delivery
        - No Stock
        - Confirmed
        - Shipped
        - Delivered
        - Cancelled
      example: Created
    Orden:
      type: object
      description: Orden de compra con los identificadores de los productos pedidos
      required: [id, email, direccionEnvio, telefono, estado, fechaCreacion, productos]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1001
        email:
          type: string
          format: email
          maxLength: 150
          example: cliente@email.com
        direccionEnvio:
          type: string
          minLength: 1
          maxLength: 250
          example: Av. Italia 3333, Maldonado
        telefono:
          type: string
          pattern: '^\+?[0-9\s\-]{7,20}$'
          example: '+59899111222'
        estado:
          $ref: '#/components/schemas/EstadoOrden'
        fechaCreacion:
          type: string
          format: date-time
          readOnly: true
          description: Fecha y hora de solicitud del pedido (ISO 8601)
          example: '2026-06-21T14:30:00-03:00'
        productos:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineaOrden'
    OrdenCreate:
      type: object
      description: |
        Datos necesarios para registrar una orden. El `id`, el `estado` y la
        `fechaCreacion` son asignados por el servidor.
      required: [email, direccionEnvio, telefono, productos]
      properties:
        email:
          type: string
          format: email
          maxLength: 150
        direccionEnvio:
          type: string
          minLength: 1
          maxLength: 250
        telefono:
          type: string
          pattern: '^\+?[0-9\s\-]{7,20}$'
        productos:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineaOrden'
    OrdenCreada:
      type: object
      description: Identificador de la orden generada
      required: [id]
      properties:
        id:
          type: integer
          format: int64
          example: 1001
    LineaOrdenDetalle:
      type: object
      description: Línea de la orden con todos los datos del producto solicitado
      required: [cantidad, producto]
      properties:
        cantidad:
          type: integer
          format: int32
          minimum: 1
          example: 2
        subtotal:
          type: number
          format: double
          description: precioUnitario x cantidad
          example: 2501.00
        producto:
          $ref: '#/components/schemas/Producto'
    OrdenDetalle:
      type: object
      description: Orden de compra con el detalle completo de cada producto pedido
      required: [id, email, direccionEnvio, telefono, estado, fechaCreacion, productos]
      properties:
        id:
          type: integer
          format: int64
          example: 1001
        email:
          type: string
          format: email
          example: cliente@email.com
        direccionEnvio:
          type: string
          example: Av. Italia 3333, Maldonado
        telefono:
          type: string
          example: '+59899111222'
        estado:
          $ref: '#/components/schemas/EstadoOrden'
        fechaCreacion:
          type: string
          format: date-time
          example: '2026-06-21T14:30:00-03:00'
        total:
          type: number
          format: double
          description: Suma de los subtotales de todas las líneas
          example: 3000.50
        productos:
          type: array
          items:
            $ref: '#/components/schemas/LineaOrdenDetalle'
    Mensaje:
      type: object
      description: Mensaje informativo de una operación exitosa
      required: [mensaje]
      properties:
        mensaje:
          type: string
          example: Producto actualizado correctamente
    Error:
      type: object
      description: Mensaje de error
      required: [codigo, mensaje]
      properties:
        codigo:
          type: integer
          format: int32
          description: Código de estado HTTP
          example: 404
        mensaje:
          type: string
          description: Descripción del error
          example: No existe el producto con id 99
        detalles:
          type: array
          description: Detalle de los errores de validación, si corresponde
          items:
            type: string
    OrdenEstadoUpdate:
      type: object
      description: Resultado del procesamiento a aplicar sobre la orden
      required: [estado]
      properties:
        estado:
          type: string
          enum:
            - Ready to Delivery
            - No Stock
          example: Ready to Delivery
    FacturaItem:
      type: object
      description: Ítem de factura, uno por cada producto solicitado
      required: [productoId, descripcion, cantidad, precioUnitario, subtotal]
      properties:
        productoId:
          type: integer
          format: int64
          example: 1
        descripcion:
          type: string
          description: Nombre del producto al momento de facturar
          example: Notebook Lenovo ThinkPad
        cantidad:
          type: integer
          format: int32
          example: 2
        precioUnitario:
          type: number
          format: double
          example: 1250.50
        subtotal:
          type: number
          format: double
          description: precioUnitario x cantidad
          example: 2501.00
    Factura:
      type: object
      description: |
        Factura de una orden que quedó en `Ready to Delivery`. El total es la suma de los
        subtotales de todos los ítems.
      required: [id, ordenId, fecha, email, direccionEnvio, telefono, items, total]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        ordenId:
          type: integer
          format: int64
          example: 1
        fecha:
          type: string
          format: date-time
          example: '2026-06-21T14:30:05-03:00'
        email:
          type: string
          format: email
          example: cliente@email.com
        direccionEnvio:
          type: string
          example: Av. Italia 3333, Maldonado
        telefono:
          type: string
          example: '+59899111222'
        items:
          type: array
          items:
            $ref: '#/components/schemas/FacturaItem'
        total:
          type: number
          format: double
          example: 3000.50
...