# ordenes/src/main/java/uy/edu/utec/taller/ordenes/dto/MensajeDTO.java
package uy.edu.utec.taller.ordenes.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class MensajeDTO {

    private String mensaje;
}